# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "version": "0.0.1",
    "title": "Pet store",
    "description": "Pet store Node.js learning project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://your-pet-server.onrender.com",
      "description": "Main server"
    }
  ],
...
  "paths": {
...
    "/api/news/": {
      "get": {
        "operationId": "getAllNews",
        "tags": [
          "News"
        ],
        "summary": "Find all news",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "number of items to skip",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "max records to return",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News successfully send",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundErrorResponse"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "HTTP status code",
            "example": "401"
          },
          "message": {
            "type": "string",
            "description": "Message to display",
            "example": "Email or password is invalid"
          },
          "ErrorName": {
            "type": "string",
            "description": "Error name",
            "example": "HttpError"
          }
        }
      },
...
      "NewsResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 200
          },
          "data": {
            "type": "object",
            "description": "news information",
            "example": {
              "imgUrl": "https://www.nytimes.com/images/2023/04/16/magazine/",
              "title": "What I Learned Dogsitting",
              "text": "In a city of yawning class...",
              "date": "2023-04-11T09:00:18+0000",
              "url": "https://www.nytimes.com/2023/04/11...",
              "id": "nyt://article/8d29f1fc-d146-509d-8ceb-5a5b17d7886b"
            }
          }
        }
      }
    },
...
    "responses": {
...
      "NotFoundErrorResponse": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "NotFound": {
                "$ref": "#/components/examples/NotFoundErrorExample"
              }
            }
          }
        }
      }
    },
    "examples": {
...
      "NotFoundErrorExample": {
        "value": {
          "code": "404",
          "message": "Not found",
          "ErrorName": "HttpError"
        }
      }
    }
  }
}